Paper reminders

+++ membership/management/commands/paper_reminders.py
@@ -10,6 +10,11 @@ from membership.reference_numbers import barcode_4
 from django.conf import settings
 
 import os
+import signal
+
+from threading import currentThread
+
+from optparse import make_option
 
 from datetime import datetime, timedelta
 from decimal import Decimal
@@ -28,10 +33,20 @@ TMPDIR = '/tmp/sikteeritex'
 class LatexTemplate(Template):
     delimiter = '\$'
 
-def get_data():
+class Timeout(Exception):
+    pass
+
+def timeout_handler(signum, frame):
+    raise Timeout
+
+def get_data(memberid=None):
     if not settings.ENABLE_REMINDERS:
         # return empty queryset
         return BillingCycle.objects.filter(id=-1)
+    elif memberid:
+        print('memberid: %s' % memberid)
+        return BillingCycle.objects.filter(membership__id=memberid
+                ).exclude(bill__type='P')
     return BillingCycle.objects.annotate(bills=Count('bill')).filter(bills__gt=2,
          is_paid__exact=False,membership__status='A',membership__id__gt=-1
          ).exclude(bill__type='P').order_by('start')
@@ -49,9 +64,18 @@ def data2pdf(data):
     target.close()
     return generate_pdf(targetfile)
 
-def create_datalist():
+def create_datalist(memberid=None):
     datalist = []
-    for cycle in get_data().all():
+    for cycle in get_data(memberid).all():
+        # check if paper reminder already sent
+        cont = False
+        for bill in cycle.bill_set.all():
+            if bill.type == 'P':
+                cont=True
+                break
+        if cont:
+            continue
+
         membercontact = cycle.membership.get_billing_contact()
         data = {
             'DATE'      : datetime.now().strftime("%d.%m.%Y"),
@@ -63,12 +87,14 @@ def create_datalist():
             'EMAIL'     : membercontact.email.replace("_", "\_"),
             'OSOITE'    : membercontact.street_address,
             'POSTI'     : "%s %s" % (membercontact.postal_code, membercontact.post_office),
-            'BARCODE'   : barcode_4("FI1657413620406956",cycle.reference_number,None,cycle.sum)
+            'BARCODE'   : barcode_4(settings.IBAN_ACCOUNT_NUMBER,cycle.reference_number,None,cycle.sum)
         }
         datalist.append(data)
     return datalist
 
 def generate_pdf(latexfile):
+    # set umask to 0077
+    oldumask = os.umask(63)
     pid = Popen(['pdflatex', '-interaction=batchmode', '-output-directory=%s' % TMPDIR, 
                 '-no-file-line-error','-halt-on-error','-output-format','pdf',latexfile])
     # wait until process ends
@@ -81,13 +107,19 @@ def generate_pdf(latexfile):
         logging.error('Error processing %s: pdflatex returncode was %s' % (latexfile, pid.returncode))
         raise RuntimeError('Error processing %s: pdflatex returncode was %s' % (latexfile, pid.returncode))
         return None
-    return latexfile.replace('.tex','.pdf')
+    # Restore previous umask
+    os.umask(oldumask)
+    pdffile = latexfile.replace('.tex','.pdf')
+    if os.path.exists(pdffile):
+        return pdffile
+    else:
+        return None
 
-def generate_reminders():
+def generate_reminders(memberid=None):
     if not settings.PAPER_REMINDER_TEMPLATE or not settings.PAPER_REMINDER_TEMPLATE.endswith('.tex'):
         raise RuntimeError('Cannot create reminders without latex template!')
     elif not os.path.exists(settings.PAPER_REMINDER_TEMPLATE):
-        raise RuntimeError('reminders template file %s does not found' % settings.PAPER_REMINDER_TEMPLATE)
+        raise RuntimeError('Reminders template file %s does not found' % settings.PAPER_REMINDER_TEMPLATE)
     if not os.path.isdir(TMPDIR) and not os.path.exists(TMPDIR):
         os.mkdir(TMPDIR)
     elif os.path.exists(TMPDIR) and not os.path.isdir(TMPDIR):
@@ -95,17 +127,22 @@ def generate_reminders():
         return
     
     filelist = []
-    for data in create_datalist():
+    for data in create_datalist(memberid):
         output = data2pdf(data)
         if output != None:
             filelist.append(output)
+    if len(filelist) == 0:
+        raise RuntimeError('No need for remainders')
     single_latex = "\\documentclass[a4paper,10pt]{letter}\n"
     single_latex += "\\usepackage{pdfpages}\n"
     single_latex += "\\begin{document}\n"
     for filename in filelist:
         single_latex += "\includepdf[pages=-]{%s}\n" % filename
     single_latex += "\\end{document}\n"
-    singlefile = os.path.join(TMPDIR, 'reminders.tex')
+    if memberid:
+        singlefile = os.path.join(TMPDIR, 'reminders_%s.tex' % memberid)
+    else:
+        singlefile = os.path.join(TMPDIR, 'reminders.tex')
     f = open(singlefile, 'w')
     f.write(single_latex)
     f.close()
@@ -115,6 +152,8 @@ def generate_reminders():
         filename = os.path.join(TMPDIR, filename)
         if filename.endswith(singlefile):
             continue
+        if 'reminder' in filename and '.pdf' in filename:
+            continue
         try:
             os.remove(filename)
         except OSError as e:
@@ -123,13 +162,35 @@ def generate_reminders():
         return None
     return singlefile
 
+def get_reminders(memberid=None):
+    if currentThread().getName() == 'MainThread':
+        signal.signal(signal.SIGALRM, timeout_handler)
+        signal.alarm(28)  # 28 sec
+    else:
+        logging.error("Cannot use signals on child thread")
+    try:
+        singlefile = generate_reminders(memberid)
+        signal.alarm(0)
+    except Timeout:
+        return
+    if singlefile:
+        f = open(singlefile, 'r')
+        return f.read()
+
+
 class Command(BaseCommand):
     args = ''
     help = 'Create paper reminders pdf'
+    option_list = BaseCommand.option_list + (
+        make_option('--member',
+            dest='member',
+            default=None,
+            help='Create pdf-reminder for user'),
+        )
 
     def handle(self, *args, **options):
         try:
-            pdffile = generate_reminders()
+            pdffile = generate_reminders(memberid=options['member'])
             if pdffile:
                 print "pdf file created: %s" % pdffile
             else:
